+++ integrations/telegram/src/kafka/kafka-adapter.service.ts
