+++ integrations/telegram/src/kafka/kafka.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { MessagesModule } from 'src/messages/messages.module';
-import { KafkaAdapterService } from './kafka-adapter/kafka-adapter.service';
-import { KafkaController } from './kafka/kafka.controller';
+import { KafkaAdapterService } from './kafka-adapter.service';
+import { KafkaController } from './kafka.controller';
 import { ConfigService } from '@nestjs/config';
 import { generateKafkaClientOptions } from 'src/kafka.configuration';
 import { ClientProxyFactory } from '@nestjs/microservices';
