+++ integrations/telegram/src/messages/index.ts
@@ -1,2 +1,2 @@
 export * from './models'
-export * from './messages/messages.service'
+export * from './messages.service'
