+++ integrations/telegram/src/messages/messages.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { MessagesService } from './messages/messages.service';
+import { MessagesService } from './messages.service';
 
 @Module({
   providers: [MessagesService],
