+++ integrations/telegram/test/chats/chats-storage.service.spec.ts
@@ -1,15 +1,15 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { MessagesService } from './messages.service';
+import { ChatsStorageService } from 'src/chats/chats-storage.service';
 
-describe('MessagesService', () => {
-  let service: MessagesService;
+describe('ChatsStorageService', () => {
+  let service: ChatsStorageService;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      providers: [MessagesService],
+      providers: [ChatsStorageService],
     }).compile();
 
-    service = module.get<MessagesService>(MessagesService);
+    service = module.get<ChatsStorageService>(ChatsStorageService);
   });
 
   it('should be defined', () => {
