+++ integrations/telegram/test/telegram/telegram.service.spec.ts
@@ -1,5 +1,5 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { TelegramService } from './telegram.service';
+import { TelegramService } from 'src/telegram/telegram.service';
 
 describe('TelegramService', () => {
   let service: TelegramService;
